Fix warnings in Crawdad for 64-bit build.
Former-commit-id: cbdd0e979ef3aa2e0741e6183b2c21c53ac5efd1

+++ pwiz_tools/Shared/Crawdad/CrawPeak.cpp
@@ -54,7 +54,7 @@ namespace crawpeaks {
      
     // begin erynes BUGBUG DEBUG
     if (!raw.empty())
-      raw_area = crawutils::area_under_curve(raw, start_idx, stop_idx);
+      raw_area = (float)crawutils::area_under_curve(raw, start_idx, stop_idx);
     // end erynes BUGBUG DEBUG
 
     //calculate_slope( raw );
